added working validator and start of validation tests
