+++ BackEndNServer/src/main/java/com/proj/validators/Validatable.java
@@ -1,6 +1,8 @@
 package com.proj.validators;
 
+import com.proj.model.users.User;
+
  public interface Validatable {
-   public abstract boolean HandleString(String element);
+   public abstract boolean ValidateStringField(User user);
    public abstract void nextValidator(Validatable validator);
 }
